fix detail on my code to pass the checker

+++ doubly_linked_lists/0-print_dlistint.c
@@ -1,16 +1,17 @@
 #include "lists.h"
 #include <stdio.h>
+#include <stddef.h>
 
 /**
 * print_dlistint - that prints all the elements of a dlistint_t list
 *
 * @h: list to print
 *
-* Return: the number of node
+* Return: the number of node(s)
 */
 size_t print_dlistint(const dlistint_t *h)
 {
-	unsigned int nbNode = 0;
+	size_t nbNode = 0;
 
 	while (h != NULL)
 	{
